+++ src/deployment/image-poll.test.ts
