+++ src/deployment/test-driver.ts
@@ -1,5 +1,5 @@
 import nock from 'nock';
-import { DeploymentDescriptor } from './deployment-descriptor';
+import {DeploymentDescriptor} from "./deployment-descriptor";
 
 export function nockDeploymentManifestJson(desc: DeploymentDescriptor, times = 1) {
   let registryScope = nock('https://orbs-network.github.io'); // expect manifest
